fixed: create beneficiery form. reworked PersonSelect and added info text for better UX in Campaign Creation page
