+++ public/locales/bg/beneficiary.json
@@ -2,11 +2,11 @@
   "all": "Всички бенефициенти",
   "grid": {
     "id": "ID",
-    "individual": "Човек",
-    "company": "Компания",
+    "individual": "Личност",
+    "company": "Юридическо лице",
     "coordinator": "Координатор",
     "countryCode": "Код на държава",
-    "not-person": "Не е човек",
+    "not-person": "Не е личност",
     "not-company": "Не е компания",
     "city": "Град",
     "coordinatorRelation": "Отношения с координатор",
@@ -24,7 +24,15 @@
   },
   "forms": {
     "add-heading": "Добави бенефициент",
-    "edit-heading": "Редактирай бенефициент"
+    "edit-heading": "Редактирай бенефициент",
+    "info": "Бенефициент може да бъде всеки потребител или юридическо лице регистрирано в платформата",
+    "labels": {
+      "type": {
+        "individual": "Личност",
+        "company": "Юридическо лице"
+      },
+      "personSelect": "Изберете потребител"
+    }
   },
   "actions": "Действия",
   "deleteContent": "Сигурни ли сте, че искате да изтриете тези запис/и?",
