+++ public/locales/bg/coordinator.json
@@ -12,6 +12,8 @@
   },
   "create": {
     "heading": "Добави нов координатор",
+    "info": "Координатор може да бъде всеки регистриран потребител в платформата",
+    "personSelect": "Изберете потребител",
     "submitButton": "Създай"
   }
 }
